Moved handlers to routes.rs
